kono-judge/requirements.txt
websockets
python-dotenv